EB, EBS updates
[EnemyBulletSystem.cs] added optional parameter to SpawnBullet() to spawn bullets not relative to enemy position
[AriesBulletSystem4.cs] re-coded attack pattern
[AriesBulletSystem41.cs] updated attack pattern
[AriesBullet1, 2, 7.cs] modified speed

+++ Assets/Scripts/Game/Bullet/Enemy bullet systems/Aries/AriesBulletSystem1.cs
@@ -14,7 +14,7 @@ public class AriesBulletSystem1 : EnemyBulletSystem
             {
                 for (int j = 0; j < 3; j++)
                 {
-                    int z = (i * 45) + (j * 120);
+                    float z = (i * 45f) + (j * 120f);
                     SpawnBullet(0, z, Vector2.zero);
                 }
 
@@ -27,7 +27,7 @@ public class AriesBulletSystem1 : EnemyBulletSystem
             {
                 for (int j = 0; j < 30; j++)
                 {
-                    int z = j * 12 + (i * 6);
+                    float z = j * 12f + (i * 6f);
                     SpawnBullet(1, z, Vector2.zero);
                 }
